refactor: toggle as common component
